Fixed function

+++ 14-binary_tree_balance.c
@@ -7,7 +7,7 @@
  */
 int binary_tree_balance(const binary_tree_t *tree)
 {
-	size_t left_heigh, right_height;
+	int left_heigh, right_height;
 
 	if (tree == NULL)
 	{
